requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-ninja==0.21.0
psycopg==3.1.9
psycopg-binary==3.1.9
pydantic==1.10.7
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
